Mejora recorte de los bloques de sonido. TODO: al exportar los cambios de longitud no se producen

+++ src/main/java/org/example/tunesfx/audio/PlaylistItem.java
@@ -4,7 +4,7 @@ public class PlaylistItem {
     private String patternName; // Qué suena
     private int startBar;       // Cuándo empieza (compás)
     private int trackIndex;     // En qué pista está
-    private int durationBars;
+    private double durationBars;
 
     public PlaylistItem(String patternName, int startBar, int trackIndex) {
         this.patternName = patternName;
@@ -19,6 +19,6 @@ public class PlaylistItem {
     // Setters (necesarios para el Drag & Drop)
     public void setStartBar(int startBar) { this.startBar = startBar; }
     public void setTrackIndex(int trackIndex) { this.trackIndex = trackIndex; }
-    public int getDurationBars() { return durationBars; }
-    public void setDurationBars(int durationBars) {this.durationBars = durationBars; }
+    public double getDurationBars() { return durationBars; }
+    public void setDurationBars(double durationBars) {this.durationBars = durationBars; }
 }
